Extract single-project removal flows into useProjectRemovalActions
The unpin-isolation fix pushed ProjectList.tsx past its 900-line
ceiling; per the guard's own advice, extract rather than raise. The
delete / remove-from-app confirm state and handlers now live beside
their bulk counterparts, sharing one unpin-isolation helper.
